Send webhook notifications through an ActiveJob
Part of the migration from Delayed::Job
