+++ app/controllers/api/v1/web_hooks_controller.rb
@@ -32,7 +32,7 @@ class Api::V1::WebHooksController < Api::BaseController
     webhook = @api_key.user.web_hooks.new(url: @url)
     @rubygem ||= Rubygem.find_by_name("gemcutter")
 
-    if webhook.fire(request.protocol.delete("://"), request.host_with_port, @rubygem,
+    if webhook.fire(request.protocol.delete("://"), request.host_with_port,
       @rubygem.versions.most_recent, delayed: false)
       render plain: webhook.deployed_message(@rubygem)
     else
