+++ test/models/web_hook_test.rb
@@ -137,43 +137,6 @@ class WebHookTest < ActiveSupport::TestCase
     end
   end
 
-  context "with a rubygem and version" do
-    setup do
-      @rubygem = create(:rubygem, name: "foogem", downloads: 42)
-      @version = create(:version,
-        rubygem: @rubygem,
-        number: "3.2.1",
-        authors: %w[AUTHORS],
-        description: "DESC")
-      @hook    = create(:web_hook, rubygem: @rubygem)
-      @job     = Notifier.new(@hook.url, "http", "localhost:1234", @rubygem, @version)
-    end
-
-    should "have gem properties encoded in JSON" do
-      payload = JSON.load(@job.payload)
-      assert_equal "foogem",    payload["name"]
-      assert_equal "3.2.1",     payload["version"]
-      assert_equal "ruby",      payload["platform"]
-      assert_equal "DESC",      payload["info"]
-      assert_equal "AUTHORS",   payload["authors"]
-      assert_equal 42,          payload["downloads"]
-      assert_equal "http://localhost:1234/gems/foogem", payload["project_uri"]
-      assert_equal "http://localhost:1234/gems/foogem-3.2.1.gem", payload["gem_uri"]
-    end
-
-    should "send the right version out even for older gems" do
-      new_version = create(:version, number: "2.0.0", rubygem: @rubygem)
-      new_hook    = create(:web_hook)
-      job         = Notifier.new(new_hook.url, "http", "localhost:1234", @rubygem, new_version)
-      payload     = JSON.load(job.payload)
-
-      assert_equal "foogem", payload["name"]
-      assert_equal "2.0.0",  payload["version"]
-      assert_equal "http://localhost:1234/gems/foogem", payload["project_uri"]
-      assert_equal "http://localhost:1234/gems/foogem-2.0.0.gem", payload["gem_uri"]
-    end
-  end
-
   context "with a non-global hook job" do
     setup do
       @url     = "http://example.com/gemcutter"
@@ -186,7 +149,7 @@ class WebHookTest < ActiveSupport::TestCase
       authorization = Digest::SHA2.hexdigest(@rubygem.name + @version.number + @hook.user.api_key)
       RestClient.expects(:post).with(anything, anything, has_entries("Authorization" => authorization))
 
-      @hook.fire("https", "rubygems.org", @rubygem, @version, delayed: false)
+      @hook.fire("https", "rubygems.org", @version)
     end
 
     should "include an Authorization header for a user with no API key" do
@@ -194,7 +157,7 @@ class WebHookTest < ActiveSupport::TestCase
       authorization = Digest::SHA2.hexdigest(@rubygem.name + @version.number)
       RestClient.expects(:post).with(anything, anything, has_entries("Authorization" => authorization))
 
-      @hook.fire("https", "rubygems.org", @rubygem, @version, delayed: false)
+      @hook.fire("https", "rubygems.org", @version)
     end
 
     should "include an Authorization header for a user with many API keys" do
@@ -203,11 +166,11 @@ class WebHookTest < ActiveSupport::TestCase
       authorization = Digest::SHA2.hexdigest(@rubygem.name + @version.number + @hook.user.api_keys.first.hashed_key)
       RestClient.expects(:post).with(anything, anything, has_entries("Authorization" => authorization))
 
-      @hook.fire("https", "rubygems.org", @rubygem, @version, delayed: false)
+      @hook.fire("https", "rubygems.org", @version)
     end
 
     should "not increment failure count for hook" do
-      @hook.fire("https", "rubygems.org", @rubygem, @version, delayed: false)
+      @hook.fire("https", "rubygems.org", @version)
 
       assert_predicate @hook.failure_count, :zero?
     end
@@ -233,7 +196,7 @@ class WebHookTest < ActiveSupport::TestCase
        Net::ProtocolError].each_with_index do |exception, index|
         RestClient.stubs(:post).raises(exception)
 
-        @hook.fire("https", "rubygems.org", @rubygem, @version, delayed: false)
+        @hook.fire("https", "rubygems.org", @version)
 
         assert_equal index + 1, @hook.reload.failure_count
         assert_predicate @hook, :global?
